:truck: Move buildLabel util-method to its own file
